--- sever/database/database_test.go
@@ -1,12 +1,0 @@
-package database
-
-import (
-	"testing"
-)
-
-func TestOpenDatabase(t *testing.T) {
-}
-
-func TestTJW(t *testing.T) {
-	
-}
